Adjusted to work with database
Setup so that ...compare/[id] where [id] is the name of the tool to compare against other tools.

Currently the page will still not be able to access the database due to requiring a school to access the tools.

Might need to add a default school or fallback tool access on the API without needing to specify a school for it to work without context.